Refactor `nannou::wgpu` module into a `nannou_wgpu` crate
This allows for more easily using nannou's various wgpu extensions in
downstream projects without needing to include the rest of `nannou`.
This might be particularly useful for headless nannou apps or libraries
that aim to extend nannou's graphical/compute capabilities.

In our case this should also be useful for developing a slimmer
`nannou_ui` crate, without requiring the entirety of the `nannou` crate
as a dependency.