Refactor `nannou_egui` and add some basic docs
This refactor aims to clean up the `nannou_egui` code a little, and
provide a slightly more modular, cleaner interface (though for the most
part usage is almost identical other than some name changes).

One notable change is the use of RAII to "end the frame". The aim is to
avoid the situation where users forget to call `end_frame`, or
accidentally call `end_frame` on the inner `CtxRef` rather than
top-level API. `begin_frame` now returns a `FrameCtx` - a type that
derefs to a `CtxRef` and automatically calls `end_frame` and tesselates
graphics when it is dropped. Users can call `FrameCtx::end` if they wish
to end the frame before `drop` is invoked. Alternatively, they can just
call `begin_frame` within a block.

@alexene let me know your thoughts! I think I'll probably start using
this in a large downstream project soon.