Ensure `wakeup` calls (`UserEvent`s) provide updates in `Wait` mode
After landing #782 I noticed that the `App` proxy's `wakeup` method
hasn't been functional since the last event loop refactor (when wgpu was
introduced).

I also noticed that #782 makes the extra updates provided in the `Wait`
loop mode redundant. Previously, we ensured at least ~3 updates would
occur following the most recently received event. This was aimed at
allowing GUI's to update for a couple of extra frames to resolve subtle
1-3 frame animations, but I'm unsure this is really the right approach.
At the very least, the number of updates should be configurable as a
part of the `LoopMode::Wait` variant, but that can wait for a future PR.