Expose `begin()` method in `geom::path::Builder`
As part of the lyon 0.17 upgrade, the `move_to` method got removed from
the builder. This makes it impossible to start a path/subpath without
having to access the underlying lyon Builder.

This commit exposes the `begin()` method in nannou's Builder, which is
what `move_to` got renamed to in lyon.

closes: #806