Bump winit to 0.27. Fix private access to window.
The commit updates nannou and nannou_egui to use winit to 0.27+. The
window attribute of the WindowBuilder is now private, we use the
WindowBuilder methods to properly build the window in window.rs.

Because the properties of the `Window` currently being built cannot be
accessed from the `WindowBuilder` anymore, some additional properties
have to be stored in the `Builder` struct which are mirrors of the
`Window`'s properties. These properties allow checking if a value and
what it was set to. `Builder` now has `fullscreen` `size`, `min_size`,
and `max_size` as new properties.

The `set_cursor_grab` has also been updated with the new winit enum
`CursorGrabMode`. If the cursor should be grabbed, we first try to
confine it, then try to lock it (see
https://docs.rs/winit/latest/winit/window/enum.CursorGrabMode.html).