Handle new events from winit 0.27.
The Ime event corresponds to a "composition event". It will never be
sent if `Window::set_ime_allowed` is not called. These events are
stateful, they first need to be enabled and successive events need to be
combined. This should be a separate feature since it will require some
extra work to handle.

The occluded event is handled a function can be used to take advantage
of it. For example, costly rendering can be avoided if the window is
occluded (hidden from view).