Remove Cargo.lock from gitignore. Include to ensure working examples.
The nannou has a large number of examples and a few standalone binaries
whose reproducibility could benefit by including the lock file in the
repo.

While it's generally recommended to gitignore the Cargo.lock for
libraries, it can also result in spurious breakage that can add to the
maintenance burden. As we tend to run `cargo update` with each version
release anyway, including the Cargo.lock in git tracking should not
cause any major issues.